requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
PyJWT==2.8.0
